# app/api/links/[id]/up/route.ts
import { NextResponse } from "next/server";
import { db } from "@/lib/db";
import { verifyWebhookToken } from "@/lib/webhook";
import { writeAudit, extractIp } from "@/lib/audit";

async function handleUp(req: Request, id: string) {
  const url = new URL(req.url);
  const token =
    req.headers.get("x-webhook-token") ?? url.searchParams.get("token") ?? "";

  if (!verifyWebhookToken(id, token)) {
    return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
  }

  const link = await db.link.findUnique({ where: { id } });
  if (!link) return NextResponse.json({ error: "Link not found" }, { status: 404 });

  const now = new Date();

  if (!link.isOnline) {
    await db.$transaction([
      db.linkEvent.create({ data: { linkId: id, type: "UP", timestamp: now } }),
      db.link.update({ where: { id }, data: { isOnline: true, lastEventAt: now } }),
    ]);
  }

  // SEC-022: audit webhook events with source IP
  void writeAudit({
    action: "UPDATE", entity: "Link", entityId: id, entityName: link.name,
    details: { event: "up", previousStatus: link.isOnline ? "up" : "down" },
    ipAddress: extractIp(req), userId: null, username: "webhook",
  });

  return NextResponse.json({ ok: true, status: "up" });
}

export async function POST(req: Request, { params }: { params: Promise<{ id: string }> }) {
  const { id } = await params;
  return handleUp(req, id);
}

export async function GET(req: Request, { params }: { params: Promise<{ id: string }> }) {
  const { id } = await params;
  return handleUp(req, id);
}