# docs/openapi.yaml
openapi: 3.1.0
info:
  title: WatchIT Tower Monitoramento — API
  version: "1.0.0"
  description: |
    API REST do WatchIT Tower Monitoramento — dashboard local de monitoramento de infraestrutura de TI.

    ## Autenticação

    A maioria dos endpoints exige uma sessão NextAuth ativa (cookie de sessão JWT).
    Faça login em `POST /api/auth/callback/credentials` e inclua o cookie `next-auth.session-token`
    nas requisições subsequentes.

    ### Exceções (sem autenticação de sessão)
    - `GET /api/health` — público, sem autenticação
    - `POST /api/links/{id}/up` e `POST /api/links/{id}/down` — autenticados por HMAC-SHA256 via header `x-webhook-token` ou query string `?token=`

    ## Rate Limiting
    O endpoint de login (`POST /api/auth/callback/credentials`) tem rate limit de
    **10 tentativas por IP em 15 minutos**.
servers:
  - url: http://localhost:3000
    description: Desenvolvimento local
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: Cookie de sessão JWT gerado pelo NextAuth após login bem-sucedido.

    webhookToken:
      type: apiKey
      in: header
      name: x-webhook-token
      description: |
        Token HMAC-SHA256 gerado com `WEBHOOK_SECRET + linkId`.
        Pode ser enviado como header `x-webhook-token` ou query string `?token=`.
  schemas:
...
    DeviceType:
      type: string
      enum: [MIKROTIK, DVR, CAMERA, OTHER, UNIFI_AP, OMADA_AP]
      description: Tipo do dispositivo.
...
    Device:
      type: object
      description: Dispositivo monitorado com todas as suas configurações de protocolo.
      properties:
        id:
          type: string
          description: CUID único do dispositivo.
          example: "clxyz123456789"
        name:
          type: string
          example: "Mikrotik Matriz"
        ip:
          type: string
          format: ipv4
          example: "192.168.1.1"
        type:
          $ref: "#/components/schemas/DeviceType"
        location:
          type: string
          nullable: true
          example: "Rack Sala de TI"
        notes:
          type: string
          nullable: true
          example: "Roteador principal. Não reiniciar sem aviso."
        pingEnabled:
          type: boolean
          default: true
        httpEnabled:
          type: boolean
          default: false
        httpPort:
          type: integer
          nullable: true
          minimum: 1
          maximum: 65535
          example: 80
        httpPath:
          type: string
          default: "/"
          example: "/status"
        snmpEnabled:
          type: boolean
          default: false
        snmpCommunity:
          type: string
          default: "public"
        snmpPort:
          type: integer
          default: 161
        routerosEnabled:
          type: boolean
          default: false
        routerosUser:
          type: string
          nullable: true
          example: "admin"
        routerosPass:
          type: string
          nullable: true
          writeOnly: true
          example: "senha123"
        routerosPort:
          type: integer
          default: 8728
        unifiEnabled:
          type: boolean
          default: false
          description: Se o monitoramento UniFi está habilitado (apenas para tipo UNIFI_AP).
        unifiAuthMethod:
          type: string
          nullable: true
          enum: [apikey, userpass]
          description: |
            Método de autenticação no controlador UniFi.
            - `apikey` — Integration API (X-API-KEY); não expõe RSSI/SSID por cliente
            - `userpass` — Inform API (usuário/senha); expõe RSSI, SSID e canal por cliente
        unifiApiKey:
          type: string
          nullable: true
          writeOnly: true
          description: Chave de API para autenticação via Integration API (nunca retornada na leitura).
        unifiUser:
          type: string
          nullable: true
          writeOnly: true
          description: Usuário para autenticação via Inform API (nunca retornado na leitura).
        unifiPass:
          type: string
          nullable: true
          writeOnly: true
          description: Senha para autenticação via Inform API (nunca retornada na leitura).
        hasUnifiCredentials:
          type: boolean
          readOnly: true
          description: Indica se credenciais UniFi estão armazenadas (substitui os campos de credencial na leitura).
        unifiControllerIp:
          type: string
          nullable: true
          format: ipv4
          example: "192.168.1.200"
          description: IP do controlador UniFi (quando diferente do IP do AP).
        unifiPort:
          type: integer
          nullable: true
          default: 443
          description: Porta HTTPS do controlador UniFi.
        unifiSite:
          type: string
          nullable: true
          default: "default"
          description: Nome ou ID do site no controlador UniFi.
        unifiTlsVerify:
          type: boolean
          nullable: true
          default: false
          description: Se deve validar o certificado TLS do controlador (desabilite para certificados autoassinados).
        omadaEnabled:
          type: boolean
          default: false
          description: Se o monitoramento Omada está habilitado (apenas para tipo OMADA_AP).
        omadaClientId:
          type: string
          nullable: true
          writeOnly: true
          description: OAuth2 client_id da Northbound API do Omada (nunca retornado na leitura).
        omadaClientSecret:
          type: string
          nullable: true
          writeOnly: true
          description: OAuth2 client_secret da Northbound API do Omada (nunca retornado na leitura).
        hasOmadaCredentials:
          type: boolean
          readOnly: true
          description: Indica se credenciais Omada estão armazenadas (substitui os campos de credencial na leitura).
        omadacId:
          type: string
          nullable: true
          description: omadacId do controller Omada (descoberto via API de sites).
          example: "abc123def456"
        omadaSite:
          type: string
          nullable: true
          description: Nome do site Omada (apenas para exibição).
          example: "Sede"
        omadaSiteId:
          type: string
          nullable: true
          description: ID do site Omada (usado diretamente nas chamadas de API).
          example: "site-abc123"
        omadaControllerIp:
          type: string
          nullable: true
          format: ipv4
          example: "10.0.0.19"
          description: IP do Omada Controller (quando diferente do IP do AP).
        omadaTlsVerify:
          type: boolean
          default: true
          description: Se deve validar o certificado TLS do controller (desabilite para certificados autoassinados).
        alertWebhookUrl:
          type: string
          format: uri
          nullable: true
          description: URL para receber alertas quando o dispositivo atinge `alertThreshold` falhas consecutivas. Suporta Slack, Teams, Telegram e qualquer endpoint HTTP/HTTPS.
          example: "https://hooks.slack.com/services/T00/B00/xxx"
        alertThreshold:
          type: integer
          minimum: 1
          maximum: 100
          default: 3
          description: Número de falhas consecutivas necessárias para disparar o alerta. Cooldown de 1 hora entre disparos.
        lastAlertAt:
          type: string
          format: date-time
          nullable: true
          readOnly: true
          description: Timestamp do último alerta enviado (usado para controle de cooldown).
        checkInterval:
          type: integer
          description: Intervalo entre checagens em segundos (mínimo 10, máximo 3600).
          default: 60
          minimum: 10
          maximum: 3600
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        currentStatus:
          $ref: "#/components/schemas/DeviceStatus"
          nullable: true
      required: [id, name, ip, type, pingEnabled, httpEnabled, snmpEnabled, routerosEnabled, checkInterval, createdAt, updatedAt]
    DeviceStatus:
      type: object
      description: Resultado mais recente da checagem do dispositivo (upsertado a cada ciclo).
      properties:
        id:
          type: string
        deviceId:
          type: string
        isOnline:
          type: boolean
          description: Se o dispositivo respondeu ao último check.
        pingMs:
          type: integer
          nullable: true
          description: Latência do ping em milissegundos.
          example: 3
        httpOk:
          type: boolean
          nullable: true
          description: Se o HTTP check retornou código 2xx.
        uptime:
          type: integer
          nullable: true
          description: Uptime do dispositivo em segundos (via SNMP ou RouterOS).
        cpuLoad:
          type: number
          format: float
          nullable: true
          description: Uso de CPU em percentual (0-100).
          example: 12.5
        memoryUsed:
          type: number
          format: float
          nullable: true
          description: Uso de memória em percentual (0-100).
          example: 45.2
        checkedAt:
          type: string
          format: date-time
          description: Timestamp da última checagem.
        unifiData:
          $ref: "#/components/schemas/UnifiResult"
          nullable: true
          description: Dados do AP UniFi (presente apenas para dispositivos do tipo UNIFI_AP).
        omadaData:
          $ref: "#/components/schemas/OmadaResult"
          nullable: true
          description: Dados do AP Omada (presente apenas para dispositivos do tipo OMADA_AP).
      required: [id, deviceId, isOnline, checkedAt]
...
    DeviceInput:
      type: object
      description: Campos para criar ou atualizar um dispositivo. Todos os campos são opcionais no PUT (PATCH semântico).
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          example: "Mikrotik Filial Norte"
        ip:
          type: string
          format: ipv4
          example: "10.0.1.1"
          description: Endereço IPv4 do dispositivo (formato `x.x.x.x`).
        type:
          $ref: "#/components/schemas/DeviceType"
        location:
          type: string
          maxLength: 200
          example: "Rack Filial Norte"
        notes:
          type: string
          maxLength: 1000
        pingEnabled:
          type: boolean
          default: true
        httpEnabled:
          type: boolean
          default: false
        httpPort:
          type: integer
          minimum: 1
          maximum: 65535
          nullable: true
        httpPath:
          type: string
          default: "/"
        snmpEnabled:
          type: boolean
          default: false
        snmpCommunity:
          type: string
          default: "public"
        snmpPort:
          type: integer
          default: 161
        routerosEnabled:
          type: boolean
          default: false
        routerosUser:
          type: string
          nullable: true
          deprecated: true
          description: Deprecated — use credenciais criptografadas. Campo aceito para compatibilidade mas nunca retornado pela API.
        routerosPass:
          type: string
          nullable: true
          deprecated: true
          description: Deprecated — use credenciais criptografadas. Campo aceito para compatibilidade mas nunca retornado pela API.
        routerosPort:
          type: integer
          default: 8728
        checkInterval:
          type: integer
          minimum: 10
          maximum: 3600
          default: 60
        alertWebhookUrl:
          type: string
          format: uri
          nullable: true
          description: URL de webhook para receber alertas de falha (Slack, Teams, HTTP genérico). Disparado após `alertThreshold` falhas consecutivas com cooldown de 1 hora.
          example: "https://hooks.slack.com/services/T00/B00/xxx"
        alertThreshold:
          type: integer
          minimum: 1
          maximum: 100
          default: 3
          description: Número de falhas consecutivas para disparar o alerta.
...
    ErrorResponse:
      type: object
      properties:
        error:
          oneOf:
            - type: string
              example: "Unauthorized"
            - type: object
              description: Objeto de erros de validação Zod (fieldErrors + formErrors).
      required: [error]
    WebhookResponse:
      type: object
      properties:
        ok:
          type: boolean
          example: true
        status:
          type: string
          enum: [up, down]
          example: "up"
      required: [ok, status]

# ─── Security global (aplicada a todos os endpoints, sobrescrita individualmente) ───
security:
  - sessionCookie: []

# ─── Paths ──────────────────────────────────────────────────────────────────
paths:
...
  /api/devices:
    get:
      tags: [Dispositivos]
      summary: Listar dispositivos
      description: |
        Retorna todos os dispositivos com o status mais recente (`currentStatus`) incluído.
        Use o parâmetro `type` para filtrar por tipo — útil ao popular dropdowns de Mikrotik em formulários de link.

        **Paginação**: quando `page` ou `limit` são informados, a resposta inclui o header `X-Total-Count` com o total de registros.
        Sem paginação, retorna todos os registros.

        **Cache / ETag**: quando `type` não é usado, a resposta inclui um header `ETag` calculado via `COUNT + MAX(updatedAt)`.
        O cliente pode enviar `If-None-Match` para receber `304 Not Modified` quando os dados não mudaram.
      parameters:
        - in: query
          name: type
          schema:
            $ref: "#/components/schemas/DeviceType"
          description: Filtrar por tipo de dispositivo.
          example: MIKROTIK
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
          description: "Número da página (base 1). Requer `limit`."
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
          description: Registros por página. Ativa a paginação e o header `X-Total-Count`.
      responses:
        "200":
          description: Lista de dispositivos ordenada por nome.
          headers:
            ETag:
              description: Fingerprint do conjunto de dados para cache condicional (presente quando `type` não é usado).
              schema:
                type: string
              example: '"a1b2c3d4e5f6"'
            X-Total-Count:
              description: Total de registros (presente quando `page` ou `limit` são usados).
              schema:
                type: integer
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Device"
        "400":
          description: Tipo de dispositivo inválido.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "Tipo de dispositivo inválido"
        "401":
          $ref: "#/components/responses/Unauthorized"

    post:
      tags: [Dispositivos]
      summary: Criar dispositivo
      description: Cria um novo dispositivo monitorado. O worker começará a monitorá-lo no próximo ciclo.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeviceInput"
            example:
              name: "Mikrotik Matriz"
              ip: "192.168.1.1"
              type: "MIKROTIK"
              location: "Sala de TI"
              pingEnabled: true
              routerosEnabled: true
              routerosUser: "admin"
              routerosPass: "senha123"
              checkInterval: 30
      responses:
        "201":
          description: Dispositivo criado com sucesso.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /api/links/{id}/up:
    parameters:
      - $ref: "#/components/parameters/LinkId"
    post:
      tags: [Links]
      summary: Webhook — marcar link como UP
      description: |
        Registra um evento UP e atualiza `isOnline = true` do link.
        Se o link já estiver online, a chamada é idempotente (retorna `ok: true` sem criar evento duplicado).

        **Autenticação:** token HMAC-SHA256 via header `x-webhook-token` ou query string `?token=`.
        O token é gerado com `HMAC-SHA256(WEBHOOK_SECRET, linkId)`.

        Aceita tanto `POST` quanto `GET` para compatibilidade com scripts e ferramentas de monitoramento externas.
      security:
        - webhookToken: []
      parameters:
        - in: query
          name: token
          schema:
            type: string
          description: Token de autenticação (alternativa ao header `x-webhook-token`).
      responses:
        "200":
          description: Link marcado como UP.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookResponse"
              example:
                ok: true
                status: "up"
        "401":
          description: Token inválido ou ausente.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          $ref: "#/components/responses/NotFound"
    get:
      tags: [Links]
      summary: Webhook — marcar link como UP (GET)
      description: Mesmo comportamento do `POST`. Suportado para compatibilidade com ferramentas que só fazem GET.
      security:
        - webhookToken: []
      parameters:
        - in: query
          name: token
          schema:
            type: string
      responses:
        "200":
          description: Link marcado como UP.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookResponse"
        "401":
          description: Token inválido ou ausente.
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/devices/test-omada:
    post:
      tags: [Dispositivos]
      summary: Testar conexão Omada e listar sites
      description: |
        Valida as credenciais OAuth2 da Northbound API do Omada Controller e retorna a lista de
        sites disponíveis. Use antes de salvar um OMADA_AP para confirmar o `omadacId` e selecionar
        o `omadaSiteId` correto.

        **Resolução de credenciais:**
        - Se `omadaClientId` e `omadaClientSecret` forem enviados, são usados diretamente.
        - Se omitidos mas `deviceId` for fornecido, as credenciais são lidas do banco (descriptografadas).
        - Se nenhuma credencial estiver disponível, retorna 422.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [controllerIp, omadacId, tlsVerify]
              properties:
                controllerIp:
                  type: string
                  format: ipv4
                  description: IP do Omada Controller.
                  example: "10.0.0.19"
                omadacId:
                  type: string
                  description: omadacId do controller (visível na URL do painel Omada).
                  example: "abc123def456"
                tlsVerify:
                  type: boolean
                  description: Se deve validar o certificado TLS (false para certificados autoassinados).
                  example: false
                omadaClientId:
                  type: string
                  description: OAuth2 client_id (opcional se deviceId for fornecido).
                omadaClientSecret:
                  type: string
                  description: OAuth2 client_secret (opcional se deviceId for fornecido).
                deviceId:
                  type: string
                  description: ID de um OMADA_AP existente para reutilizar as credenciais armazenadas.
      responses:
        "200":
          description: Autenticação bem-sucedida. Retorna lista de sites.
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    example: true
                  sites:
                    type: array
                    items:
                      type: object
                      properties:
                        siteId:
                          type: string
                          example: "site-abc123"
                        name:
                          type: string
                          example: "Sede"
                  message:
                    type: string
                    example: "Autenticado. 2 site(s) disponível(is)."
        "400":
          description: Parâmetros inválidos.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          description: |
            Credenciais ausentes ou falha na conexão/autenticação com o controller.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                sem_credenciais:
                  summary: Credenciais não fornecidas
                  value:
                    error: "Informe client_id e client_secret para testar a conexão"
                auth_falhou:
                  summary: client_id ou client_secret inválidos
                  value:
                    error: "client_id ou client_secret inválidos"
                conexao_falhou:
                  summary: Controller inacessível
                  value:
                    error: "Timeout ao conectar a 10.0.0.19"
...
  /api/reports:
    get:
      tags: [Status]
      summary: Relatório de disponibilidade por dispositivo
      description: |
        Retorna um relatório detalhado para um ou mais dispositivos com:
        - **Resumo:** uptime%, total de checks, contagem de incidentes, downtime total, ping médio/min/max
        - **Insights:** alertas automáticos (uptime crítico, ping elevado, etc.)
        - **Histórico de ping:** série temporal de `isOnline` e `pingMs`
        - **Histórico RouterOS:** série temporal de CPU e memória (apenas para MIKROTIK com RouterOS habilitado)
        - **Incidentes:** períodos de indisponibilidade com duração

        Máximo de 10 dispositivos por requisição.
      parameters:
        - in: query
          name: devices
          required: true
          schema:
            type: string
          description: IDs dos dispositivos separados por vírgula.
          example: "clxyz123,clxyz456"
        - in: query
          name: hours
          schema:
            type: integer
            default: 168
            minimum: 1
            maximum: 720
          description: Janela de tempo em horas (máximo 720h = 30 dias). Padrão 168h (7 dias).
      responses:
        "200":
          description: Relatórios gerados com sucesso (um por dispositivo).
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    device:
                      type: object
                      properties:
                        id:
                          type: string
                        name:
                          type: string
                        type:
                          $ref: "#/components/schemas/DeviceType"
                        ip:
                          type: string
                        location:
                          type: string
                          nullable: true
                    period:
                      type: object
                      properties:
                        from:
                          type: string
                          format: date-time
                        to:
                          type: string
                          format: date-time
                        hours:
                          type: integer
                    summary:
                      type: object
                      properties:
                        uptimePct:
                          type: number
                          format: float
                          description: Percentual de uptime (0–100).
                          example: 99.31
                        totalChecks:
                          type: integer
                          example: 10080
                        onlineChecks:
                          type: integer
                          example: 10010
                        incidentCount:
                          type: integer
                          example: 2
                        totalDowntimeMs:
                          type: integer
                          description: Tempo total offline em milissegundos.
                          example: 4200000
                        avgPingMs:
                          type: number
                          nullable: true
                          example: 3.2
                        maxPingMs:
                          type: number
                          nullable: true
                          example: 210
                        minPingMs:
                          type: number
                          nullable: true
                          example: 1
                    insights:
                      type: array
                      items:
                        type: object
                        properties:
                          level:
                            type: string
                            enum: [ok, warn, critical]
                          text:
                            type: string
                            example: "Uptime crítico: 72.3% nas últimas 168h"
                    pingHistory:
                      type: array
                      items:
                        type: object
                        properties:
                          timestamp:
                            type: string
                            format: date-time
                          pingMs:
                            type: integer
                            nullable: true
                          isOnline:
                            type: boolean
                    routerosHistory:
                      type: array
                      nullable: true
                      items:
                        type: object
                        properties:
                          timestamp:
                            type: string
                            format: date-time
                          cpuLoad:
                            type: number
                            nullable: true
                          memoryUsed:
                            type: number
                            nullable: true
                    incidents:
                      type: array
                      items:
                        type: object
                        properties:
                          startAt:
                            type: string
                            format: date-time
                          endAt:
                            type: string
                            format: date-time
                            nullable: true
                          durationMs:
                            type: integer
                            nullable: true
                          resolved:
                            type: boolean
        "400":
          description: Parâmetros inválidos (nenhum dispositivo informado ou limite de 10 excedido).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ── Live traffic ──────────────────────────────────────────────────────────
...